+++ InventoryManagement.Data/Models/ExpenseTransportation.cs
@@ -14,7 +14,7 @@ namespace InventoryManagement.Data
         public int VoucherNo { get; set; }
         public bool IsApproved { get; set; }
         public int RegistrationId { get; set; }
-        public double TotalExpense { get; set; }
+        public decimal TotalExpense { get; set; }
         public string ExpenseNote { get; set; }
         public DateTime ExpenseDate { get; set; }
         public DateTime InsertDate { get; set; }
