+++ InventoryManagement.Data/Models/PurchasePaymentList.cs
@@ -8,7 +8,7 @@ namespace InventoryManagement.Data
         public int PurchasePaymentListId { get; set; }
         public int PurchasePaymentId { get; set; }
         public int PurchaseId { get; set; }
-        public double PurchasePaidAmount { get; set; }
+        public decimal PurchasePaidAmount { get; set; }
         public DateTime InsertDate { get; set; }
 
         public virtual Purchase Purchase { get; set; }
