+++ InventoryManagement.Data/Models/Selling.cs
@@ -16,26 +16,26 @@ namespace InventoryManagement.Data
         public int RegistrationId { get; set; }
         public int CustomerId { get; set; }
         public int SellingSn { get; set; }
-        public double SellingTotalPrice { get; set; }
-        public double SellingDiscountAmount { get; set; }
-        public double SellingDiscountPercentage { get; set; }
-        public double SellingPaidAmount { get; set; }
-        public double SellingReturnAmount { get; set; }
-        public double SellingDueAmount { get; set; }
+        public decimal SellingTotalPrice { get; set; }
+        public decimal SellingDiscountAmount { get; set; }
+        public decimal SellingDiscountPercentage { get; set; }
+        public decimal SellingPaidAmount { get; set; }
+        public decimal SellingReturnAmount { get; set; }
+        public decimal SellingDueAmount { get; set; }
         public string SellingPaymentStatus { get; set; }
         public DateTime SellingDate { get; set; }
         public DateTime? LastUpdateDate { get; set; }
         public DateTime InsertDate { get; set; }
 
         public DateTime? PromisedPaymentDate { get; set; }
-        public double ServiceCharge { get; set; }
-        public double ServiceCost { get; set; }
+        public decimal ServiceCharge { get; set; }
+        public decimal ServiceCost { get; set; }
         //[ServiceCharge]-[ServiceCost]
-        public double ServiceProfit { get; set; }
+        public decimal ServiceProfit { get; set; }
         public string ServiceChargeDescription { get; set; }
-        public double Expense { get; set; }
+        public decimal Expense { get; set; }
         public string ExpenseDescription { get; set; }
-        public double BuyingTotalPrice { get; set; }
+        public decimal BuyingTotalPrice { get; set; }
         public decimal SellingAccountCost { get; set; }
         //[BuyingTotalPrice]-([SellingTotalPrice]+[SellingDiscountAmount]+[Expense]+[SellingAccountCost])
         public decimal SellingProfit { get; set; }
