+++ InventoryManagement.Data/Models/ServicePaymentList.cs
@@ -8,7 +8,7 @@ namespace InventoryManagement.Data
         public int ServicePaymentListId { get; set; }
         public int SellingPaymentId { get; set; }
         public int ServiceId { get; set; }
-        public double ServicePaidAmount { get; set; }
+        public decimal ServicePaidAmount { get; set; }
         public DateTime InsertDate { get; set; }
 
         public virtual SellingPayment SellingPayment { get; set; }
