+++ InventoryManagement.Repository/Repositories/Customer/CustomerViewModel.cs
@@ -16,7 +16,7 @@ namespace InventoryManagement.Repository
         public string CustomerName { get; set; }
         public string CustomerAddress { get; set; }
         public string Description { get; set; }
-        public double DueLimit { get; set; }
+        public decimal DueLimit { get; set; }
         public string Designation { get; set; }
         public bool IsIndividual { get; set; }
         public byte[] Photo { get; set; }
@@ -39,8 +39,8 @@ namespace InventoryManagement.Repository
         public string CustomerName { get; set; }
         public string CustomerAddress { get; set; }
         public string PhonePrimary { get; set; }
-        public double Due { get; set; }
-        public double DueLimit { get; set; }
+        public decimal Due { get; set; }
+        public decimal DueLimit { get; set; }
         public string Description { get; set; }
         public string Designation { get; set; }
         public bool IsIndividual { get; set; }
@@ -65,26 +65,26 @@ namespace InventoryManagement.Repository
         public string CustomerName { get; set; }
         public string CustomerAddress { get; set; }
         public string Description { get; set; }
-        public double DueLimit { get; set; }
+        public decimal DueLimit { get; set; }
         public string Designation { get; set; }
         public bool IsIndividual { get; set; }
         public byte[] Photo { get; set; }
         public DateTime SignUpDate { get; set; }
         public ICollection<CustomerPhoneViewModel> PhoneNumbers { get; set; }
         public ICollection<CustomerSellingViewModel> SellingRecords { get; set; }
-        public double SoldAmount { get; set; }
-        public double ReceivedAmount { get; set; }
-        public double DiscountAmount { get; set; }
-        public double DueAmount { get; set; }
+        public decimal SoldAmount { get; set; }
+        public decimal ReceivedAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
+        public decimal DueAmount { get; set; }
     }
     public class CustomerSellingViewModel
     {
         public int SellingId { get; set; }
         public int SellingSn { get; set; }
-        public double SellingAmount { get; set; }
-        public double SellingPaidAmount { get; set; }
-        public double SellingDiscountAmount { get; set; }
-        public double SellingDueAmount { get; set; }
+        public decimal SellingAmount { get; set; }
+        public decimal SellingPaidAmount { get; set; }
+        public decimal SellingDiscountAmount { get; set; }
+        public decimal SellingDueAmount { get; set; }
         public DateTime SellingDate { get; set; }
     }
 
@@ -92,6 +92,6 @@ namespace InventoryManagement.Repository
     {
         public int CustomerId { get; set; }
         public string Name { get; set; }
-        public double Due { get; set; }
+        public decimal Due { get; set; }
     }
 }
