+++ InventoryManagement.Repository/Repositories/ExpenseFixed/ExpenseFixedModel.cs
@@ -3,7 +3,7 @@
     public class ExpenseFixedAddModel
     {
         public string Name { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public int IntervalDays { get; set; }
     }
 
@@ -11,8 +11,8 @@
     {
         public int ExpenseFixedId { get; set; }
         public string Name { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public int IntervalDays { get; set; }
-        public double CostPerDay { get; set; }
+        public decimal CostPerDay { get; set; }
     }
 }
