+++ InventoryManagement.Repository/Repositories/PurchasePayment/PurchasePaymentViewModel.cs
@@ -10,7 +10,7 @@ namespace InventoryManagement.Repository
     public class PurchasePaymentListViewModel
     {
         public string PaymentMethod { get; set; }
-        public double PaidAmount { get; set; }
+        public decimal PaidAmount { get; set; }
         public DateTime PaidDate { get; set; }
     }
 
@@ -20,7 +20,7 @@ namespace InventoryManagement.Repository
         public int VendorId { get; set; }
         public int RegistrationId { get; set; }
         public string PaymentMethod { get; set; }
-        public double PaidAmount { get; set; }
+        public decimal PaidAmount { get; set; }
         public DateTime PaidDate { get; set; }
     }
 }
