+++ InventoryManagement.Repository/ViewModel/Expense/ExpenseAllViewModel.cs
@@ -10,7 +10,7 @@ namespace InventoryManagement.Repository
         public bool IsTransportation { get; set; }
         public string CreateBy { get; set; }
         public string ExpenseCategory { get; set; }
-        public double ExpenseAmount { get; set; }
+        public decimal ExpenseAmount { get; set; }
         public string ExpenseFor { get; set; }
         public DateTime ExpenseDate { get; set; }
     }
